chore(flutter): reorganize code
