+++ flutter/measure_flutter/example/integration_test/plugin_integration_test.dart
@@ -1,15 +1,5 @@
-// This is a basic Flutter integration test.
-//
-// Since integration tests run in a full Flutter application, they can interact
-// with the host side of a plugin implementation, unlike Dart unit tests.
-//
-// For more information about Flutter integration tests, please see
-// https://flutter.dev/to/integration-testing
-
-
 import 'package:flutter_test/flutter_test.dart';
 import 'package:integration_test/integration_test.dart';
-
 import 'package:measure_flutter/measure_flutter.dart';
 
 void main() {
