+++ flutter/measure_flutter/test/measure_flutter_method_channel_test.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/services.dart';
 import 'package:flutter_test/flutter_test.dart';
-import 'package:measure_flutter/measure_flutter_method_channel.dart';
+import 'package:measure_flutter/src/measure_flutter_method_channel.dart';
 
 void main() {
   TestWidgetsFlutterBinding.ensureInitialized();
